ENH: Improved lapack, ATLAS, and OpenBLAS support
CMake was improved to allow for choosing the vendor
of the lapack libraries desired.

NOTE:  LAPACKE distributions prior to 3.6.0 often reference lapack test
functions that are not distributed with lapack!  This often causes linkage problems
similar to:
NOTE:  LAPACKE 3.5.0 on openSUSE has linkage problems with

/usr/lib64/gcc/x86_64-suse-linux/4.8/../../../../lib64/liblapacke.so: undefined reference to clatms_
... <CLIPPED> ...
/usr/lib64/gcc/x86_64-suse-linux/4.8/../../../../lib64/liblapacke.so: undefined reference to zlatms_

This is fixed in lapack 3.6.0 !
https://bugzilla.redhat.com/show_bug.cgi?id=1287405