COMP: Add Stage 2 build for bart binaries
The 2nd stage of the Makefile build was not implemented
in cmake.  When this shortcomming was identified, it was
better understood that each source file is compiled multiple
times with different flags for different purposes.

The rather verbose approach taken to preserve backwards
compatibility, and to avoid multiple conditional compilation
with different flags was to use the source files as templates
that are re-written to the build tree, and then compiled.

This approach facilitates debugging (the code compiled can
be viewed from the build tree, without identiftying the
result of the pre-processor.