png: add more png_write functions
Since png_write was only used in bart toimg, it only supported the one
format used by toimg. For a new image export tool in view
(https://github.com/mrirecon/view), it is very useful to also use bart's
libpng interface. However, view uses the cairo graphics library, which
has a different convention on the byte order in RGB data. Therefore, we
need a BGR version of png_write. Since it also uses a 4-byte buffer per
pixel, we need a BGR32 (or BGRA) version.

This commits adds a generic function, which can deal with all of those
cases, as well as 4 wrappers. One of the wrappers corresponds to the
previous function.