cuda fft: fix behavior and add unit test (#210)
This commit fixes an unfortunate bug with the
CUDA fft interface, cuFFT:
if a transform is along several, non-adjacent dimensions
in such a way that it skips some dimensions, the cuFFT
will fail. Unfortunately, not always with an error,
sometimes it will silently generate a plan which does
not work correclty.

Therefore, we now detect whether the flags are non-adjacent
and genrate chained transforms in those cases.

Additionally, this commit adds a unit test which checks such cases.

This fixes the bug in SMS-NLINV on the GPU reported by Pim Borman.