misc: unlink already opened files
In order to properly support identical input and output in all our
tools, we now remove the input file if it is later opened for writing.

This necessitates that we always create the output files
after all inputs, but this is the logical order anyway.

Since we have a private memory-mapping to the input, we can remove
the file from the filesystem and overwrite it with the output.

Caveat: This might not work on Windows.