properly support inoutfiles
Additionally, it is now an error to open a file for either reading or
writing without first registering its use as an output, input, or inout
file. With declared OPT{IN,OUT}FILE and ARG_{IN,OUT}FILE options, there
should be no need now to open a file without an earlier declaration of
intended use.