improve parsing of int, long and float
atoi, atol and family hide errors by returning 0 when they cannot parse
their string argument. This can lead to bugs when not giving an argument
to e.g. OPT_INT:
	bart nlinv -i -d4
would set the number of iteration (-i) to zero, as it cannot parse -d4
as an integer.

Now, such cases lead to an error.