fix bart opts parsing
We now stop parsing bart options at the first unknown option.
For this, we need to prepent '+' to the optstr passed to ya_getopt,
which needs a new argument to our `options()`, `bool stop_at_nonopt`.

Additionally, `parse_bart_opts()` now prints the correct usage string
(generated from the bart options), and changes argv and argc to remove
the options it parsed.