Makefile: clear out some default variables by hand
In a single-stage Makefile, we cannot us `-rR` to clear out all implicit
rules and variables, as the variables are defined before reading in the
Makefile, so before make gets to our MAKEFLAGS += ...

So we just undefine the variables which we might conceivably overwrite.
We need to take care to only undefine them if they come from their
default values, otherwise syntax such as `CC=clang make` will not work
anymore.
The full list of builtin variables if far too long to include here.