Makefile: allclean: remove commands even from top-level dir
When upgrading bart repositories from before the move to `commands/`, it
can happen that, e.g. avg still exists in the top level dir.
However, since we specify dependencies in the tests as just "avg" and
add `commands/` to the vpath, our automatic tests would pick up `avg` in
the top-level dir and therefore not compile it, but then try to call
`commands/avg` which does not exist.

Therefore, we are not also removing any leftover files with the same
names as our commands in the top-level dir