WASM: fix unmapping of .ra files
For .ra files, we called `munmp()` with the wrong size, since we
truncate the pointer (to get the pointer of the mapping without the
header), but did not add that difference to the size.

If we do add that difference, we can call bart tools using node.js
running through wasm with .ra files, which means that we can run our
tests!