Makefile: Correct version check for AR_LOCK_NEEDED
Turns out it is in 4.4.1, which will only be easily checkable once GNU
Make 4.5 releases.
So to not get any nasty surprises, we comment out the entire check for
now, but let people set AR_LOCK_NEEDED=0 by hand if they know that their
version of GNU Make is new enough.