fix copy tool, improve stream_sync_slice.
1) The copy tool has a stream-specific feature, --streams, which allows
creating streams from a fixed data set. This was subtly broken:
If the output was a stream (because "-" was specified as an output
file), but `--streams` was not given, and the input was a stream
but with a larger flag, e.g. 1024, then the copy loop would wrongly
sync the whole output file albeit when only one slice from the input
stream had been copied.

2) fixed problematic behavior of stream_sync_slice, which would sync
an output stream slice if only part of the respective slice was specified.